requirements.lambda.txt
fastapi
uvicorn
opencv-python-headless
numpy
python-dotenv
pydantic-ai-slim[openai,google]
python-multipart
httpx
mangum
awslambdaric
